Move new measurement functions into the measure source files
